+++ modules/planning/tuning/speed_model/autotuning_speed_mlp_model.h
@@ -55,11 +55,11 @@ class AutotuningSpeedMLPModel : public AutotuningBaseModel {
    * @param flat_feature [eigen matrix, row: time, col: traj point feature]
    */
   void FlattenFeatures(const autotuning::TrajectoryFeature& feature,
-                       Eigen::MatrixXf* const flat_feature) const;
+                       Eigen::MatrixXd* const flat_feature) const;
 
   void FlattenFeatures(
       const autotuning::SpeedPointwiseFeature& speed_point_feature,
-      const int row, Eigen::MatrixXf* const flat_feature) const;
+      const int row, Eigen::MatrixXd* const flat_feature) const;
 };
 
 }  // namespace planning
